refactor User into /data
