+++ app/src/androidTest/java/com/github/fribourgsdp/radio/BuggyLobbyActivityTest.kt
@@ -12,6 +12,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.mockimplementations.BuggyLobbyActivity
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
