+++ app/src/androidTest/java/com/github/fribourgsdp/radio/FirestoreDatabaseTest.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.DocumentSnapshot
