+++ app/src/androidTest/java/com/github/fribourgsdp/radio/GameTest.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import org.junit.Test
 
 import org.junit.Assert.*
