+++ app/src/androidTest/java/com/github/fribourgsdp/radio/HostGameHandlerTest.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.view.View
 import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.mockimplementations.FakeGameView
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentSnapshot
