+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/LocalDatabase.kt
@@ -3,6 +3,7 @@ package com.github.fribourgsdp.radio.mockimplementations
 import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentSnapshot
