+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockGameSettingsActivity.kt
@@ -5,6 +5,7 @@ import android.os.Bundle
 import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import org.mockito.Mockito.*
 
 class MockGameSettingsActivity: GameSettingsActivity() {
