+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockJoinGameActivities.kt
@@ -8,7 +8,7 @@ import com.github.fribourgsdp.radio.JoinGameActivity
 import com.github.fribourgsdp.radio.JoinWithQRCodeFragment
 import com.github.fribourgsdp.radio.mockimplementations.BuggyDatabase
 import com.github.fribourgsdp.radio.mockimplementations.LocalDatabase
-import com.github.fribourgsdp.radio.User
+import com.github.fribourgsdp.radio.data.User
 import org.mockito.Mockito.*
 
 class WorkingJoinGameActivity : JoinGameActivity() {
