+++ app/src/main/java/com/github/fribourgsdp/radio/FireBaseTestActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import com.google.android.gms.tasks.Tasks
 
 class FireBaseTestActivity : MyAppCompatActivity() {
