+++ app/src/main/java/com/github/fribourgsdp/radio/FirestoreDatabase.kt
@@ -4,6 +4,7 @@ import android.content.ContentValues
 import android.util.Log
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.Timestamp
