+++ app/src/main/java/com/github/fribourgsdp/radio/Game.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import java.util.*
 import kotlin.collections.ArrayList
 import kotlin.collections.HashMap
