+++ app/src/main/java/com/github/fribourgsdp/radio/GameSettingsActivity.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import android.view.View
 import android.widget.*
 import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.User
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
