+++ app/src/main/java/com/github/fribourgsdp/radio/ImportSpotifyPlaylistsActivity.kt
@@ -5,6 +5,7 @@ import android.content.Intent
 import android.os.Bundle
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import okhttp3.*
 import org.json.JSONException
 import java.io.IOException
