+++ app/src/main/java/com/github/fribourgsdp/radio/PlayerGameHandler.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.content.Intent
 import android.util.Log
+import com.github.fribourgsdp.radio.data.User
 import com.google.firebase.firestore.DocumentSnapshot
 
 class PlayerGameHandler(
