+++ app/src/main/java/com/github/fribourgsdp/radio/PlaylistSongsFragment.kt
@@ -9,6 +9,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
