+++ app/src/main/java/com/github/fribourgsdp/radio/SongFragment.kt
@@ -8,6 +8,7 @@ import android.widget.EditText
 import android.widget.TextView
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.data.User
 
 open class SongFragment : MyFragment(R.layout.fragment_song) {
     private lateinit var initialLyrics : String
