+++ app/src/main/java/com/github/fribourgsdp/radio/UserPlaylistsFragment.kt
@@ -6,6 +6,7 @@ import android.view.View
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.User
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
